+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/ast/impl/javacc/AbstractTokenManager.java
@@ -12,8 +12,8 @@ import java.util.Map;
 
 import net.sourceforge.pmd.PMDConfiguration;
 import net.sourceforge.pmd.lang.TokenManager;
-import net.sourceforge.pmd.lang.ast.AstInfo.SuppressionCommentWrapper;
-import net.sourceforge.pmd.lang.ast.AstInfo.SuppressionCommentWrapper.SuppressionCommentImpl;
+import net.sourceforge.pmd.lang.ast.impl.SuppressionCommentImpl;
+import net.sourceforge.pmd.reporting.ViolationSuppressor.SuppressionCommentWrapper;
 
 /**
  * A base class for the token managers generated by JavaCC.
