+++ pmd-java/src/main/javacc/Java.jjt
@@ -333,8 +333,8 @@ import java.util.EnumSet;
 import java.util.List;
 import java.util.Set;
 import java.util.Map;
-import net.sourceforge.pmd.lang.ast.AstInfo.SuppressionCommentWrapper;
-import net.sourceforge.pmd.lang.ast.AstInfo.SuppressionCommentWrapper.SuppressionCommentImpl;
+import net.sourceforge.pmd.reporting.ViolationSuppressor.SuppressionCommentWrapper;
+import net.sourceforge.pmd.lang.ast.impl.SuppressionCommentImpl;
 import net.sourceforge.pmd.lang.ast.impl.javacc.JavaccToken;
 import net.sourceforge.pmd.lang.document.TextRegion;
 import net.sourceforge.pmd.lang.ast.Node;
